hjälp

+++ src/gameManager.rs
@@ -64,16 +64,16 @@ pub struct GameManager{
 impl GameManager{
     pub fn new() -> Self {
         let gameParameters = HashMap::from([
-            ("towerAttackDamage", 10.0),
+            ("towerAttackDamage", 1.0),
             ("enemyAttackDamage", 1.0),
             ("enemyHealth", 50.0),
             ("baseHealth", 10.0),
             ("enemyHeight", 1.0),
             ("towerHeight", 0.3),
-            ("enemySpeed", 1.5),
+            ("enemySpeed", 1.0),
             ("projectileSpeed", 10.0),
             ("towerAttackRate", 2.0),  // How many seconds between each attack
-            ("finalWave", 10.0),
+            ("finalWave", 4.0),
             ("currentWave", 1.0),
         ]);
         Self {
@@ -118,8 +118,8 @@ impl GameManager{
 
     pub fn gameloop(&mut self){
 
-        let eye = Point3::new(10.0f32, 10.0, 10.0);
-        let at = Point3::origin();
+        let eye = Point3::new(110.0f32, 10.0, 110.0);
+        let at = Point3::new(100.0f32, 1.0, 100.0);
         let mut first_person = FirstPerson::new(eye, at);
 
         // thread for towers to attack
@@ -232,8 +232,8 @@ impl GameManager{
                 self.window.draw_text("You lose! :(", &pos, 250.0, &font, &Point3::new(0.0, 0.0, 0.4));
             }
 
-            // self.window.render_with_camera(&mut first_person);
-            self.window.render();
+            self.window.render_with_camera(&mut first_person);
+            // self.window.render();
         }
     }
 
@@ -245,6 +245,7 @@ impl GameManager{
         let collider1 = collision.collider1();
         let collider2 = collision.collider2();
         let mut types: Vec<(TypeEnum, &mut IdComponent,&mut AttackDamageComponent, &mut ColliderComponent, &mut RenderableComponent)> = Vec::new();
+        let mut actualTypes: Vec<&mut TypeComponent> = Vec::new();
 
         let mut colliderCompList = self.entityManager.borrowComponentVecMut::<ColliderComponent>().unwrap();
         let mut idCompList = self.entityManager.borrowComponentVecMut::<IdComponent>().unwrap();
@@ -256,18 +257,12 @@ impl GameManager{
         let iter = zip.filter_map(|(colliderComp, (idComp, (typeComp, (damageComp, renderComp)))),
                                                                     |Some((colliderComp.as_mut()?, idComp.as_mut()?, typeComp.as_mut()?, damageComp.as_mut()?, renderComp.as_mut()?)));
             
-        // let zip = colliderCompList.iter_mut().zip(idCompList.iter_mut().zip(typeCompList.iter_mut().zip(damageCompList.iter_mut())));
-        // let iter = zip.filter_map(|(colliderComp, (idComp, (typeComp, damageComp))),
-        //                                                                     |Some((colliderComp.as_mut()?, idComp.as_mut()?, typeComp.as_mut()?, damageComp.as_mut()?)));
         /* Loop through all objects and if it's an enemy then move it */
         for (colliderComp, idComp, typeComp, damageComp, renderComp) in iter{
-            if colliderComp.getColliderHandle().0 == collider1.0{ 
+            if colliderComp.getColliderHandle().0 == collider1.0 || colliderComp.getColliderHandle().0 == collider2.0{ 
                 types.push((typeComp.getType(), idComp, damageComp, colliderComp, renderComp));
+                actualTypes.push(typeComp);
             }
-            else if colliderComp.getColliderHandle().0 == collider2.0{ 
-                types.push((typeComp.getType(), idComp, damageComp, colliderComp, renderComp));
-            }
-            
         }
 
         let obj1 = types.pop();
@@ -279,10 +274,14 @@ impl GameManager{
             let obj1 = obj1.unwrap();
             let obj2 = obj2.unwrap();
 
-            println!("For id={} the position collider is: {} ", obj1.1.getId(), self.physicsManager.getCollider(obj1.3.getColliderHandle()).translation());
-            // let temp = self.physicsManager.getCollider(obj1.3.getColliderHandle()).parent().unwrap();
-            // println!("For id={} the position rigidbody is: {} ", obj1.1.getId(), self.physicsManager.getRigidBody(temp).unwrap().translation());
-            // println!("for id={} the position of scenenode is: {} ", obj1.1.getId(), obj1.4.getSceneNode().read().unwrap().data().local_translation());
+            let a = actualTypes.pop().unwrap();
+            let b = actualTypes.pop().unwrap();
+            // println!("obj1 type: {} - obj2 type: {}", a.getTypeString(obj1.0), b.getTypeString(obj2.0));
+
+            // println!("For id={} and type= {}, the position collider is: {} ", obj2.1.getId(), b.getTypeString(obj2.0), self.physicsManager.getCollider(obj2.3.getColliderHandle()).translation());
+            // let temp = self.physicsManager.getCollider(obj2.3.getColliderHandle()).parent().unwrap();
+            // println!("For id={} the position rigidbody is: {} ", obj2.1.getId(), self.physicsManager.getRigidBody(temp).unwrap().translation());
+            // println!("for id={} and type= {}, the position of scenenode is: {} ", obj2.1.getId(), b.getTypeString(obj2.0), obj2.4.getSceneNode().read().unwrap().data().local_translation());
            
             if matches!(obj1.0, TypeEnum::enemyType){
                 if matches!(obj2.0, TypeEnum::projectileType){
@@ -297,7 +296,7 @@ impl GameManager{
                 }
             }
     
-            if matches!(obj1.0, TypeEnum::projectileType){
+            else if matches!(obj1.0, TypeEnum::projectileType){
                 if matches!(obj2.0, TypeEnum::enemyType){
                     println!("projectile & enemy");
                     self.eventManager.sendEvent(EventEnum::takeDamageEvent { id: obj2.1.getId(), damage: obj1.2.getAttackDamage() as usize});
@@ -305,12 +304,11 @@ impl GameManager{
                 }
                 else if matches!(obj2.0, TypeEnum::wallType){
                     println!("projectile & wall");
-                    //println!("For id={} the position collider is: {} ", obj2.1.getId(), self.physicsManager.getCollider(obj2.3.getColliderHandle()).translation());
                     self.eventManager.sendEvent(EventEnum::removeObjectEvent { id: obj1.1.getId(), colliderHandle: obj1.3.getColliderHandle()});
                 }
             }
     
-            if matches!(obj1.0, TypeEnum::baseType){
+            else if matches!(obj1.0, TypeEnum::baseType){
                 if matches!(obj2.0, TypeEnum::enemyType){
                     println!("base & enemy");
                     self.eventManager.sendEvent(EventEnum::takeDamageEvent { id: obj1.1.getId(), damage: obj2.2.getAttackDamage() as usize});
@@ -318,7 +316,7 @@ impl GameManager{
                 }
             }
     
-            if matches!(obj1.0, TypeEnum::wallType){
+            else if matches!(obj1.0, TypeEnum::wallType){
                 if matches!(obj2.0, TypeEnum::projectileType){
                     println!("wall & projectile");
                     self.eventManager.sendEvent(EventEnum::removeObjectEvent { id: obj2.1.getId(), colliderHandle: obj2.3.getColliderHandle()});
@@ -353,8 +351,10 @@ impl GameManager{
             // if no more waves:
             let currentWave = *self.gameParameters.get("currentWave").unwrap();
             let finalWave = *self.gameParameters.get("finalWave").unwrap();
+            println!("currentWave: {}", currentWave);
+            println!("finalWave: {}", finalWave);
 
-            if currentWave == finalWave {
+            if currentWave > finalWave {
                 self.gameState = GameStateEnum::won;
             }
             else {
@@ -401,9 +401,21 @@ impl GameManager{
         }
 
         if let EventEnum::removeObjectEvent{id, colliderHandle} = event {
-            self.entityManager.removeObject(id);
-            self.physicsManager.removeRigidBodyWithCollider(colliderHandle);
-            // self.window.remove_node(node);
+            let mut idExists = false;
+            let mut idCompList = self.entityManager.borrowComponentVecMut::<IdComponent>().unwrap();
+            let iter = idCompList.iter_mut().filter_map(|idComp| Some(idComp.as_mut()?));
+            for idComp in iter{
+                if idComp.getId() == id{
+                    idExists = true;
+                }
+            }
+            drop(idCompList);
+            if idExists == true{
+                self.entityManager.removeObject(id);
+                self.physicsManager.removeRigidBodyWithCollider(colliderHandle);
+                // self.window.remove_node(node);
+            }
+
         }
 
         // Create the necessary components for a tower
@@ -454,7 +466,15 @@ impl GameManager{
                 // Sets the renderableComponent node coordinates to the rigidBody coordinates
                 node.write().unwrap().set_local_translation(Translation3::new(t.0, t.1, t.2));
                 node.write().unwrap().set_local_rotation(r);
-                self.physicsManager.getCollider(colliderComp.getColliderHandle()).set_translation(vector![t.0, t.1, t.2])
+                // self.physicsManager.getCollider(colliderComp.getColliderHandle()).set_translation(vector![t.0, t.1, t.2]);
+
+
+                println!("The position of collider is: {} ", self.physicsManager.getCollider(colliderComp.getColliderHandle()).translation());
+                // let temp = self.physicsManager.getCollider(obj2.3.getColliderHandle()).parent().unwrap();
+                // println!("For id={} the position rigidbody is: {} ", obj2.1.getId(), self.physicsManager.getRigidBody(temp).unwrap().translation());
+                println!("The position of scenenode is: {}, {}, {} ", t.0, t.1, t.2);
+
+
             }
             if matches!(typeComp.getType(), TypeEnum::projectileType){
                 let node = renderComp.getSceneNode();
@@ -511,8 +531,7 @@ impl GameManager{
         self.entityManager.addComponentToObject(base, AttackDamageComponent::new(0.0));
 
         self.createRenderComponent(base, TypeEnum::baseType, endCoords.0, 0.2, endCoords.1, 0.007);
-        self.createRigidBodyAndColliderComponents(base, TypeEnum::baseType, endCoords.0, 0.2, endCoords.1, (0.0, 0.0, 0.0), 1.0);
-
+        self.createRigidBodyAndColliderComponents(base, TypeEnum::baseType, endCoords.0, 0.2, endCoords.1, (0.0, 0.0, 0.0), ColliderShape::cuboid(1.5, 10.0, 1.5));
     }
 
 
@@ -532,7 +551,7 @@ impl GameManager{
 
         self.createRenderComponent(projectile, TypeEnum::projectileType,  xOrigin, yOrigin, zOrigin, 0.06);
         self.createRigidBodyAndColliderComponents(projectile, TypeEnum::projectileType, 
-                     xOrigin, yOrigin, zOrigin, (xVelocity, yVelocity, zVelocity), 0.1);
+                     xOrigin, yOrigin, zOrigin, (xVelocity, yVelocity, zVelocity), ColliderShape::ball(0.3));
     }
 
 
@@ -540,7 +559,7 @@ impl GameManager{
         let tower = self.entityManager.newObject();
         self.entityManager.addComponentToObject(tower, TypeComponent::new(TypeEnum::towerType));
         self.entityManager.addComponentToObject(tower, AttackDamageComponent::new(*self.gameParameters.get("towerAttackDamage").unwrap()));
-        self.entityManager.addComponentToObject(tower, AttackRateComponent::new(*self.gameParameters.get("towerAttackRate").unwrap()));    // Do we even use this??
+        self.entityManager.addComponentToObject(tower, AttackRateComponent::new(*self.gameParameters.get("towerAttackRate").unwrap()));    
 
         self.createRenderComponent(tower, TypeEnum::towerType, x, y, z, 0.6);
     }
@@ -557,7 +576,7 @@ impl GameManager{
         let startCoords = self.mapManager.getStart();
         self.createRenderComponent(enemy, TypeEnum::enemyType, startCoords.0, *self.gameParameters.get("enemyHeight").unwrap(), startCoords.1, 0.5);
         self.createRigidBodyAndColliderComponents(enemy, TypeEnum::enemyType,
-                     startCoords.0, *self.gameParameters.get("enemyHeight").unwrap(), startCoords.1, (0.0, 0.0, 0.0), 0.5);
+                     startCoords.0, *self.gameParameters.get("enemyHeight").unwrap(), startCoords.1, (0.0, 0.0, 0.0), ColliderShape::ball(0.5));
     }
 
 
@@ -591,7 +610,7 @@ impl GameManager{
     }
 
 
-    fn createRigidBodyAndColliderComponents(&mut self, id: usize, objectType: TypeEnum, x: f32, y: f32, z: f32, velocity: (f32, f32, f32), radius: f32){
+    fn createRigidBodyAndColliderComponents(&mut self, id: usize, objectType: TypeEnum, x: f32, y: f32, z: f32, velocity: (f32, f32, f32), colliderShape: ColliderShape){
         // Add RigidBody to PhysicsManager and RigidBodyHandle to RigidBodyComponent (like an index) with a translation 
         let body = RigidBodyBuilder::new(RigidBodyType::Dynamic);
         let mut body = body.translation(vector![x, y, z]).build();
@@ -605,8 +624,9 @@ impl GameManager{
         self.entityManager.addComponentToObject(id, RigidBodyComponent::new(rigidBodyHandle));
 
         // Add Collider to PhysicsManager and ColliderHandle to ColliderComponent (like an index) with a translation 
-        let collider = ColliderBuilder::new(ColliderShape::ball(0.5));
+        let collider = ColliderBuilder::new(colliderShape);
         let collider = collider.translation(vector![x, y, z]).sensor(true).active_events(ActiveEvents::COLLISION_EVENTS).build();
+        // let colliderHandle = self.physicsManager.addCollider(collider);
         let colliderHandle = self.physicsManager.addColliderWithParent(collider, rigidBodyHandle);
         self.entityManager.addComponentToObject(id, ColliderComponent::new(colliderHandle));
     }
